Add length and padding keyworks to DistributedSampler
Current implementation of `DistributedSampler` is ideal for distributed
training using map datasets, as they fit in memory and have known size.
However, it doesn't support distributed training using `IterableDataset` datasets,
as these classes do not implement `__len__`. To fix that, a `length`
keyword was added to `DistributedSampler` and it has precedence when
set.

An extra `padding=True` parameter was added was give finer control on
whether the index lists should be padded by the sampler. This is
useful for preventing duplicate reading on `IterableDataset` datasets
that do not fit in memory or which data reading or transformation are expensive.

The re is no back compatibility with this change.